% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ButyrateProducers1}
\alias{ButyrateProducers1}
\title{ButyrateProducers1 dataframe}
\format{dataframe}
\usage{
ButyrateProducers1
}
\description{
Table of information describing the behaviour of the microbial functional group.
See help(MFG) or ?MFG for explanation of the contents of the microbial functional groups dataframes
}
\seealso{
MFG
}
\keyword{data}
\keyword{datasets}

