% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkResInfo.R
\name{checkResInfo}
\alias{checkResInfo}
\title{Checks whether the resource is included in the system information file (e.g. start value, washout rate etc)}
\usage{
checkResInfo(resNames, sys.data)
}
\arguments{
\item{resNames}{Vector of strings which contains the names of the resources in the system}

\item{sys.data}{data frame sysInfoRes i.e. resource sys info data frame}
}
\value{
nothing
}
\description{
Checks whether the resource is included in the system information file (e.g. start value, washout rate etc)
}

