% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSolution.R
\name{checkSolution}
\alias{checkSolution}
\title{Checks whether the solution generated by the ODE solver contains negative values}
\usage{
checkSolution(stateVar, out, tol = -0.1)
}
\arguments{
\item{stateVar}{Name of state variable}

\item{out}{output from microPopModel}

\item{tol}{tolerance}
}
\description{
Checks whether the solution generated by the ODE solver contains negative values
}

