% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFlowsToMoles.R
\name{convertFlowsToMoles}
\alias{convertFlowsToMoles}
\title{convertFlowsToMoles}
\usage{
convertFlowsToMoles(allStrainNames, flow, molarMass)
}
\arguments{
\item{allStrainNames}{is a vector containing the names of the microbes (strings)}

\item{flow}{is the list output from reshapeFlowMat()}

\item{molarMass}{is a named vector containing the molar mass for each resource e.g. out$parms$molarMass}
}
\description{
convert network flows from mass to moles
}
