% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertStatesToMoles.R
\name{convertStatesToMoles}
\alias{convertStatesToMoles}
\title{convertStatesToMoles}
\usage{
convertStatesToMoles(nodeMass, MolarMass)
}
\arguments{
\item{nodeMass}{is the value of each node in the network (named vector)}

\item{MolarMass}{is a named vector containing the molar mass for each resource e.g. out$parms$molarMass}
}
\description{
convert network nodes from mass to moles for resources (microbes remain as mass)
}
