% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDF.R
\name{createDF}
\alias{createDF}
\title{Create a dataframe from a CSV file}
\usage{
createDF(filename)
}
\arguments{
\item{filename}{A string containing the path to the csv file}
}
\value{
A dataframe
}
\description{
Create a dataframe from a CSV file
}

