% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivsDefault.R
\name{derivsDefault}
\alias{derivsDefault}
\title{Differential Equations called by ODE solver}
\usage{
derivsDefault(t, y, parms)
}
\arguments{
\item{t}{time}

\item{y}{vector of state variables}

\item{parms}{list of parameters}
}
\description{
Differential Equations called by ODE solver
}

