% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entryRateFuncDefault.R
\name{entryRateFuncDefault}
\alias{entryRateFunc}
\alias{entryRateFuncDefault}
\title{entry Rate Function}
\usage{
entryRateFuncDefault(varName, varValue, stateVarValues, time, inflowRate, parms)
}
\arguments{
\item{varName}{(string) Name of state variable of interest (resource name or strain name)}

\item{varValue}{(scalar) value of state variable of interest}

\item{stateVarValues}{(named vector) values of all state variables}

\item{time}{(scalar) time}

\item{inflowRate}{(named vector) on inflow rates (specified in SysInfo files)}

\item{parms}{List containing all system parameters}
}
\value{
(scalar) rate of entry (quantity per unit time) for any state variable
}
\description{
Return the rate of entry to the system for any state variable
}

