% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraGrowthLimFuncDefault.R
\name{extraGrowthLimFuncDefault}
\alias{extraGrowthLimFunc}
\alias{extraGrowthLimFuncDefault}
\title{Extra Growth Limitation Function}
\usage{
extraGrowthLimFuncDefault(strainName, groupName, pathName, stateVarValues,
  stateVarNames, time, parms)
}
\arguments{
\item{strainName}{Name of strain}

\item{groupName}{Name of group}

\item{pathName}{metabolic path name e.g. 'path1'}

\item{stateVarValues}{values of all state variables at the current time step}

\item{stateVarNames}{names of all state variables}

\item{time}{time,t, in ODE solver}

\item{parms}{list of all parameters}
}
\value{
(scalar) limitation on growth (between 0 and 1)
}
\description{
Return the value of extraGrowthLim (number between 0 and 1)
}

