% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllResources.R
\name{getAllResources}
\alias{getAllResources}
\title{Makes vector of unique resource names}
\usage{
getAllResources(microbeNames)
}
\arguments{
\item{microbeNames}{Vector of strings which contains the names of the microbial groups in the system e.g. c('Bacteroides','Acetogens')}
}
\value{
vector of resource names
}
\description{
Makes vector of unique resource names
}

