% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKeyRes.R
\name{getKeyRes}
\alias{getKeyRes}
\title{Finds the name of the key resource for each path for each MFG}
\usage{
getKeyRes(microbeNames, numPaths)
}
\arguments{
\item{microbeNames}{Vector of strings which contains the names of the microbial groups in the system e.g. c('Bacteroides','Acetogens')}

\item{numPaths}{Named vector. Number of paths for each microbial group}
}
\value{
list of vectors where the names are microbeNames
}
\description{
Finds the name of the key resource for each path for each MFG
}

