% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMolarStoichiom.R
\name{getMolarStoichiom}
\alias{getMolarStoichiom}
\title{Read in the molar stoichiom values and put in array
array has form [group,resource,path]
reads in data from MFG dataframes}
\usage{
getMolarStoichiom(microbeNames, numPaths, resNames)
}
\arguments{
\item{microbeNames}{Vector of strings which contains the names of the microbial groups in the system e.g. c('Bacteroides','Acetogens')}

\item{numPaths}{Named vector. Number of paths for each microbial group}

\item{resNames}{Vector of resource names}
}
\value{
Array [gname,rname,pathname]
}
\description{
Read in the molar stoichiom values and put in array
array has form [group,resource,path]
reads in data from MFG dataframes
}

