% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNonBoostFrac.R
\name{getNonBoostFrac}
\alias{getNonBoostFrac}
\title{obtains the none boosting fraction of growth for given MFG
if there is a boosting resource}
\usage{
getNonBoostFrac(microbeNames, resourceNames, numPaths)
}
\arguments{
\item{microbeNames}{Vector of strings which contains the names of the microbial groups in the system e.g. c('Bacteroides','Acetogens')}

\item{resourceNames}{Vector of strings which contains the names of the resources in the system}

\item{numPaths}{Named vector. Number of paths for each microbial group}
}
\value{
an array with format [group,resource,path]
}
\description{
obtains the none boosting fraction of growth for given MFG
if there is a boosting resource
}

