% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNumPaths.R
\name{getNumPaths}
\alias{getNumPaths}
\title{get the number of metabolic pathways for the given group}
\usage{
getNumPaths(microbeNames)
}
\arguments{
\item{microbeNames}{Vector of strings which contains the names of the microbial groups in the system e.g. c('Bacteroides','Acetogens')}
}
\value{
a named vector of the number of paths for each group
if numPathways is not in dataframe then it is set to 1.
}
\description{
get the number of metabolic pathways for the given group
}

