% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPHcorners.R
\name{getPHcorners}
\alias{getPHcorners}
\title{get pH corners Function}
\usage{
getPHcorners(microbeNames, pHLimit)
}
\arguments{
\item{microbeNames}{(vector of strings). Names of microbes in the system}

\item{pHLimit}{(logical) Is microbial growth affected by pH?}
}
\value{
(vector) values of the pH values of the limit function i.e. where the limit is c(0,1,1,0)
}
\description{
Returns the values of the pH values of the limit function i.e. where the limit is c(0,1,1,0)
Reads these in from the microbe group dataframes
}

