% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStrainParamsFromFile.R
\name{getStrainParamsFromFile}
\alias{getStrainParamsFromFile}
\title{get strain parameter values from a csv file}
\usage{
getStrainParamsFromFile(Pmats, strainPHcorners, strainOptions)
}
\arguments{
\item{Pmats}{List of parameter matrices}

\item{strainPHcorners}{Matrix of pH corners for each strain}

\item{strainOptions}{List which is input to microPopModel}
}
\value{
(list) - first entry is new version of Pmats, second is new version of strainPHcorners
}
\description{
get strain parameter values from a csv file
}

