% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStrainPHcorners.R
\name{getStrainPHcorners}
\alias{getStrainPHcorners}
\title{get stochastically generated pH corners for each strain}
\usage{
getStrainPHcorners(microbeNames, allStrainNames, numStrains, pHcorners, pHLimit,
  strainOptions)
}
\arguments{
\item{microbeNames}{(vector of strings). Names of microbes in the system}

\item{allStrainNames}{(vector of strings)}

\item{numStrains}{Integer}

\item{pHcorners}{vector of 4 scalars definining the pH lim func}

\item{pHLimit}{(logical) Is microbial growth affected by pH?}

\item{strainOptions}{list from microPopModel inputs}
}
\value{
(matrix) values of the pH values of the limit function i.e.
where the limit is c(0,1,1,0) for each strain
}
\description{
Returns the values of the pH values of the limit function i.e. where the limit is c(0,1,1,0)
Reads these in from the microbe group dataframes
}

