% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValues.R
\name{getValues}
\alias{getValues}
\title{get quantity for all variables in 'Names'
e.g. get 'startValue' for all resources}
\usage{
getValues(sysInfoMicrobes, sysInfoRes, Names, quantity, strainNames,
  microbeNames, resourceNames, numStrains)
}
\arguments{
\item{sysInfoMicrobes}{sys info dataframe for microbes}

\item{sysInfoRes}{sys info dataframe for resources}

\item{Names}{Vector of names of microbial group or resource}

\item{quantity}{String. Name of quantity to get value for}

\item{strainNames}{Vector of strings of strain names}

\item{microbeNames}{Vector of strings which contains the names of the microbial groups in the system e.g. c('Bacteroides','Acetogens')}

\item{resourceNames}{Vector of strings which contains the names of the resources in the system}

\item{numStrains}{Scalar. Number of strains per group}
}
\description{
get quantity for all variables in 'Names'
e.g. get 'startValue' for all resources
}

