% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeInflowFromSoln.R
\name{makeInflowFromSoln}
\alias{makeInflowFromSoln}
\title{Used for running microPop with multiple compartments
Takes the solution (state of system) from the previous compartment (out$solution)
and then finds the washout rate of each state variable using removalRateFunc
to find the inflow rate to the next downstream compartment}
\usage{
makeInflowFromSoln(out)
}
\arguments{
\item{out}{output from microPopModel()}
}
\value{
matrix of flow rates (conc/time) with named columns (the same as out$solution)
}
\description{
Used for running microPop with multiple compartments
Takes the solution (state of system) from the previous compartment (out$solution)
and then finds the washout rate of each state variable using removalRateFunc
to find the inflow rate to the next downstream compartment
}

