% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNetworkMatrices.R
\name{makeNetworkMatrices}
\alias{makeNetworkMatrices}
\title{makeNetworkMatrices}
\usage{
makeNetworkMatrices(
  chosen.time,
  out,
  convertToMoles = TRUE,
  sumOverStrains = TRUE
)
}
\arguments{
\item{chosen.time}{the time you want to plot}

\item{out}{the output from microPopModel()}

\item{convertToMoles}{Logical. Default is TRUE}

\item{sumOverStrains}{Logical. Default is TRUE}
}
\description{
make links and nodes matrices for use in network plotting software
}
