% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeParamMatrixG.R
\name{makeParamMatrixG}
\alias{makeParamMatrixG}
\title{generic function which assigns values to an array called parameterName
array has form [group,resource,path]
reads in data from MFG dataframes}
\usage{
makeParamMatrixG(microbeNames, parameterName, numPaths, resData,
  microbeMolarMass, resNames)
}
\arguments{
\item{microbeNames}{Vector of strings which contains the names of the microbial groups in the system e.g. c('Bacteroides','Acetogens')}

\item{parameterName}{Name of parameter}

\item{numPaths}{Named vector. Number of paths for each microbial group}

\item{resData}{sys info dataframe for resources}

\item{microbeMolarMass}{Scalar. Molar mass of microbe e.g. 113g/mol}

\item{resNames}{Vector of strings which contains the names of the resources in the system}
}
\value{
Array for parameterName [gname,rname,pname]
}
\description{
generic function which assigns values to an array called parameterName
array has form [group,resource,path]
reads in data from MFG dataframes
}

