% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massBalanceFuncDefault.R
\name{massBalanceFuncDefault}
\alias{massBalanceFunc}
\alias{massBalanceFuncDefault}
\title{mass balance Function}
\usage{
massBalanceFuncDefault(uptake, production, growthRate, balanceTol, strainName)
}
\arguments{
\item{uptake}{Matrix (with names) where columns are resources and rows are pathways, giving uptake rate (mass/time) of given strain}

\item{production}{Matrix (with names) where columns are resources and rows are pathways, giving production rate (mass/time) of given strain}

\item{growthRate}{(vector) microbial growth rate (mass per unit time) for one strain on each metabolic pathway}

\item{balanceTol}{(scalar) Defined in microPopModel input list checkingOptions}

\item{strainName}{(string) Name of strain in ODE solver loop}
}
\description{
Doesn't return anything but prints to screen if mass does not balance after the equations for biological growth have been derived
Only run if checkMassConv is TRUE
}

