% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanTraitFunc.R
\name{meanTraitFunc}
\alias{meanTraitFunc}
\title{calculate the mean trait at the end of the model run}
\usage{
meanTraitFunc(out, trait.name, gname, resource.name, path)
}
\arguments{
\item{out}{Output from microPopModel()}

\item{trait.name}{can be 'halfSat','yield','maxGrowthRate' and 'pHtrait' or 'strainpHcorners'}

\item{gname}{name of group or microbe}

\item{resource.name}{String}

\item{path}{String}
}
\description{
calculate the mean trait at the end of the model run
}
