% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microbeSysInfo}
\alias{microbeSysInfo}
\title{microbeSysInfo}
\format{A dataframe with the row names in the itemised list below and a column for units (optional) and for each microbial functional group (MFG) in the system to be simulated.}
\usage{
microbeSysInfo
}
\description{
Data frame describing the system information for the microbial state variables
}
\details{
Each MFG has a column. The first column can be used for describing the units of each variable. This is optional and just for clarity - it is not used within microPop (note, the units column must be labelled 'units').
The data frame must contain the following rows:
\itemize{
\item startValue The value of each MFG at the start time of the simulation (e.g. units are g/l)
\item inflowRate The value of the rate of inflow of each MFG (e.g. units are g/l/d)
\item washOut The specific washout rate of each MFG (e.g. units are /d)
}
}
\keyword{data}
\keyword{datasets}

