% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microbeSysInfoHuman}
\alias{microbeSysInfoHuman}
\title{microbeSysInfoHuman dataframe}
\format{dataframe}
\usage{
microbeSysInfoHuman
}
\description{
Table of information describing the inflows, outflows, start values of each microbial group for the R script microPop/inst/DemoFiles/human*.R
See help(microbeSysInfo) or for an explanation of the contents
}
\seealso{
microbeSysInfo
}
\keyword{data}
\keyword{datasets}

