% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkDFfromMPinput.R
\name{networkDFfromMPinput}
\alias{networkDFfromMPinput}
\title{networkDFfromMPinput}
\usage{
networkDFfromMPinput(microbeNames)
}
\arguments{
\item{microbeNames}{vector of strings of the names of the microbial data frames you want to plot. These can be intrinsic data frames or loaded in by user.}
}
\value{
a list containing the edges and nodes
}
\description{
make node and edge data frames to use in visNetwork from microPop microbial data frames
}
