% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pHcentreOfMass.R
\name{pHcentreOfMass}
\alias{pHcentreOfMass}
\title{Find the pH value which is the centre of mass of the pH limitation function
(used for the pH trait)}
\usage{
pHcentreOfMass(strainName, groupName, pHLimFunc, parms)
}
\arguments{
\item{strainName}{Name of the strain}

\item{groupName}{Name of microbial group}

\item{pHLimFunc}{function specified in rateFuncs$pHLimFunc}

\item{parms}{List of all parameters}
}
\value{
pH value at centre of mass
}
\description{
Find the pH value which is the centre of mass of the pH limitation function
(used for the pH trait)
}

