% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pHFuncDefault.R
\name{pHFuncDefault}
\alias{pHFunc}
\alias{pHFuncDefault}
\title{pH Function}
\usage{
pHFuncDefault(time, parms)
}
\arguments{
\item{time}{(scalar). The current time point in the ODE solver.}

\item{parms}{List which contains all information required by the ODE solver}
}
\value{
(scalar) pH at the given time
}
\description{
Return the value of pH in pH units
}

