% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pHLimFuncDefault.R
\name{pHLimFuncDefault}
\alias{pHLimFunc}
\alias{pHLimFuncDefault}
\title{pH Limitation Function}
\usage{
pHLimFuncDefault(strainName, groupName, pH, parms)
}
\arguments{
\item{strainName}{Name of the strain that is being looped through in the ODE solver}

\item{groupName}{Name of microbial group that is being looped through in the ODE solver}

\item{pH}{(scalar). The current pH value.}

\item{parms}{List of all parameters}
}
\value{
(scalar) pH limitation (0 to 1)
}
\description{
Return the value of pHLim (must lie in interval [0,1])
}

