% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMicrobes.R
\name{plotMicrobes}
\alias{plotMicrobes}
\title{Generic plotting of microbes over time}
\usage{
plotMicrobes(
  out,
  sumOverStrains = TRUE,
  yLabel = "Concentration",
  xLabel = "Time",
  legendPosition = "topleft",
  cex.title = 1,
  cex.ax = 1,
  cex.legend = 1
)
}
\arguments{
\item{out}{output from microPopModel()}

\item{sumOverStrains}{Logical. Default=TRUE}

\item{yLabel}{String for y axis label. Default is 'Concentration'}

\item{xLabel}{String for x axis label. Default is 'Time'}

\item{legendPosition}{String. Position of legend in microbe plot, default is 'topleft'}

\item{cex.title}{Scaling for title text}

\item{cex.ax}{Scaling for axes text (labels and ticklabels)}

\item{cex.legend}{Scaling for legend text}
}
\value{
Nothing just generates a plot
}
\description{
Generic plotting of microbes over time
}
