% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickPlot.R
\name{quickPlot}
\alias{quickPlot}
\title{Generic plotting showing results of microPop}
\usage{
quickPlot(out, numR, numStrains, microbeNames, yLabel, xLabel, sumOverStrains,
  saveFig = FALSE, figType = "eps", figName = "microPopFig")
}
\arguments{
\item{out}{Output from microPopModel()}

\item{numR}{Scalar. Number of resources}

\item{numStrains}{Scalar. Number of strains per group}

\item{microbeNames}{Vector of strings which contains the names of the microbial groups in the system e.g. c('Bacteroides','Acetogens')}

\item{yLabel}{String for y axis label}

\item{xLabel}{String for x axis label}

\item{sumOverStrains}{Logical}

\item{saveFig}{Logical}

\item{figType}{String}

\item{figName}{String}
}
\value{
Nothing just generates a plot
}
\description{
Generic plotting showing results of microPop
}

