% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickPlot1.R
\name{quickPlot1}
\alias{quickPlot1}
\title{Generic plotting showing results of microPop}
\usage{
quickPlot1(
  soln,
  numR,
  numStrains,
  microbeNames,
  yLabel,
  xLabel,
  sumOverStrains,
  resourceLegendPosition = "topleft",
  microbeLegendPosition = "topleft",
  saveFig = FALSE,
  figType = "eps",
  figName = "microPopFig"
)
}
\arguments{
\item{soln}{ODE output from microPopModel() i.e. matrix out$solution}

\item{numR}{Scalar. Number of resources}

\item{numStrains}{Scalar. Number of strains per group}

\item{microbeNames}{Vector of strings which contains the names of the microbial groups in the system e.g. c('Bacteroides','Acetogens')}

\item{yLabel}{String for y axis label}

\item{xLabel}{String for x axis label}

\item{sumOverStrains}{Logical}

\item{resourceLegendPosition}{String. Position of legend in resource plot, default is 'topleft'}

\item{microbeLegendPosition}{String. Position of legend in microbe plot, default is 'topleft'}

\item{saveFig}{Logical. Default is FALSE}

\item{figType}{String. Default is "eps"}

\item{figName}{String. Default is "microPopFig"}
}
\value{
Nothing just generates a plot
}
\description{
Generic plotting showing results of microPop
}
