% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removalRateFuncDefault.R
\name{removalRateFuncDefault}
\alias{removalRateFunc}
\alias{removalRateFuncDefault}
\title{Removal Rate Function}
\usage{
removalRateFuncDefault(varName, varValue, stateVarValues, time, washOut, parms)
}
\arguments{
\item{varName}{(string) Name of state variable of interest (this is group name or a resource name - NOT a strain name)}

\item{varValue}{(scalar) value of state variable of interest}

\item{stateVarValues}{(named vector) values of all state variables}

\item{time}{(scalar) time}

\item{washOut}{(named vector) of wash out rates (per unit time) of groups and resources (specified in SysInfo files)}

\item{parms}{List containing all system parameters}
}
\value{
(scalar) rate of removal (quantity per unit time) for the state variable varName
}
\description{
Return the rate of removal of any state variable from the system
}

