% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceListItems.R
\name{replaceListItems}
\alias{replaceListItems}
\title{used to replace items in list.in in list.default
needed for processing microPop input args like plotOptions}
\usage{
replaceListItems(list.in, list.default)
}
\arguments{
\item{list.in}{input List}

\item{list.default}{Default List}
}
\value{
list.default updated with entries from list.in
}
\description{
used to replace items in list.in in list.default
needed for processing microPop input args like plotOptions
}

