% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshapeFlowMat.R
\name{reshapeFlowMat}
\alias{reshapeFlowMat}
\title{reshapeFlowMat}
\usage{
reshapeFlowMat(time.step, flow.direction, out)
}
\arguments{
\item{time.step}{is the index of the chosen time}

\item{flow.direction}{is either 'uptake' or 'production'}

\item{out}{is the output from microPopModel with networkAnalysis=TRUE}
}
\value{
a list with microbeNames as elements and a matrix of [path,resource] showing the chosen flow direction (eg. uptake or production). Note theses flows have not been converted to moles.
}
\description{
reshapes the flow matrices out$flow.uptake or out$flow.production into a list
elements of the list are the microbeNames and then there is a matrix [path,res]
}
