% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resourceSysInfoHuman}
\alias{resourceSysInfoHuman}
\title{resourceSysInfoHuman dataframe}
\format{dataframe}
\usage{
resourceSysInfoHuman
}
\description{
Table of information describing the inflows, outflows, start values and molar masses of each resource for the R script microPop/inst/DemoFiles/human*.R
See help(resourceSysInfo) or for an explanation of the contents
}
\seealso{
resourceSysInfo
}
\keyword{data}
\keyword{datasets}

