% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resourceSysInfoRumen}
\alias{resourceSysInfoRumen}
\title{resourceSysInfoRumen dataframe}
\format{dataframe}
\usage{
resourceSysInfoRumen
}
\description{
Table of information describing the inflows, outflows, start values and molar masses of each resource for the R script microPop/inst/DemoFiles/rumen*.R
See help(resourceSysInfo) or for an explanation of the contents
}
\seealso{
resourceSysInfo
}
\keyword{data}
\keyword{datasets}

