% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMicroPopExample.R
\name{runMicroPopExample}
\alias{runMicroPopExample}
\title{runMicroPopExample}
\usage{
runMicroPopExample(name = NULL)
}
\arguments{
\item{name}{Name of the example to run. If Name is NULL the list of examples will be printed.}
}
\description{
This function is similar to the demo() function but requires less interaction
It is used to run the canned examples from the microPop package.
}

