% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{strainParams}
\alias{strainParams}
\title{strainParams dataframe}
\format{dataframe}
\usage{
strainParams
}
\description{
Table containing some parameter values for specific strains for the R script microPop/inst/DemoFiles/human4.R
The file must have colnames c(strainName, paramName, paramVal, paramUnit, resource,path) where strainName is in format 'groupName.i' where i is the strain number.
}
\keyword{data}
\keyword{datasets}

