% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumConcOverStrains.R
\name{sumConcOverStrains}
\alias{sumConcOverStrains}
\title{sumConcOverStrains}
\usage{
sumConcOverStrains(
  concentration.orig,
  allStrainNames,
  groupNames,
  resourceNames
)
}
\arguments{
\item{concentration.orig}{the row of out$solution at the required time point}

\item{allStrainNames}{is a vector containing the names of the microbial strains (strings)}

\item{groupNames}{is a vector containing the names of the microbial groups (strings)}

\item{resourceNames}{is a vector of strings containing the names of all the resources}
}
\description{
sum concentration of each strain into the group it is in
}
