% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumFlowOverStrains.R
\name{sumFlowOverStrains}
\alias{sumFlowOverStrains}
\title{sumFlowOverStrains}
\usage{
sumFlowOverStrains(flowList, allStrainNames, groupNames)
}
\arguments{
\item{flowList}{is list containing the production or uptake flows (the output from reshapeFlowMat())}

\item{allStrainNames}{is a vector containing the names of the microbial strains (strings)}

\item{groupNames}{is a vector containing the names of the microbial groups (strings)}
}
\description{
make links and nodes matrices for use in network plotting software
}
