% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumFlowsOverPaths.R
\name{sumFlowsOverPaths}
\alias{sumFlowsOverPaths}
\title{sumFlowsOverPaths}
\usage{
sumFlowsOverPaths(links)
}
\arguments{
\item{links}{data frame or matrix of links}
}
\value{
matrix of links
}
\description{
sum flows over links between the same nodes
i.e. if the link has more than one metabolic path
}
