% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{systemInfoMicrobesPhyto}
\alias{systemInfoMicrobesPhyto}
\title{systemInfoMicrobesPhyto dataframe}
\format{dataframe}
\usage{
systemInfoMicrobesPhyto
}
\description{
Table of information describing the inflows, outflows, start values of each microbial group for the R script microPop/inst/DemoFiles/phyto.R
See help(microbeSysInfo) or for an explanation of the contents
}
\seealso{
microbeSysInfo
}
\keyword{data}
\keyword{datasets}

