% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignNAsToMFGs.R
\name{assignNAsToMFGS}
\alias{assignNAsToMFGS}
\title{Assigns NA to the MFG data frames for parameter values that are not used
Even though these values are not used it is important that they are NAs when assigning random strain values
It alters the global variables that are the MFG dataframes}
\usage{
assignNAsToMFGS(microbeNames, numPaths, keyRes, resourceNames)
}
\arguments{
\item{microbeNames}{Vector of strings which contains the names of the microbial groups in the system e.g. c('Bacteroides','Acetogens')}

\item{numPaths}{Named vector. Number of paths for each microbial group}

\item{keyRes}{Name of the key resource}

\item{resourceNames}{Vector of strings which contains the names of the resources in the system}
}
\value{
nothing
}
\description{
Assigns NA to the MFG data frames for parameter values that are not used
Even though these values are not used it is important that they are NAs when assigning random strain values
It alters the global variables that are the MFG dataframes
}

