% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetFunc.R
\name{subsetFunc}
\alias{subsetFunc}
\title{this function is needed due to R dropping the names when it subsets}
\usage{
subsetFunc(array.in, gname, path)
}
\arguments{
\item{array.in}{Original array}

\item{gname}{Group name (string)}

\item{path}{Path name (string)}
}
\value{
named vector or array
}
\description{
this function is needed due to R dropping the names when it subsets
}
\keyword{internal}

